fix env path after first run

+++ build.ps1
@@ -96,14 +96,14 @@ function GetProxyEnabledWebClient {
 function Remove-PathVariable([string]$VariableToRemove) {
     $path = [Environment]::GetEnvironmentVariable("PATH", "User")
     if ($null -ne $path) {
-        $newItems = $path.Split(";", [StringSplitOptions]::RemoveEmptyEntries) | Where-Object { "$($_)" -inotlike $VariableToRemove }
-        [Environment]::SetEnvironmentVariable("PATH", [System.String]::Join(";", $newItems), "User")
+        $newItems = $path.Split($PathSplitChar, [StringSplitOptions]::RemoveEmptyEntries) | Where-Object { "$($_)" -inotlike $VariableToRemove }
+        [Environment]::SetEnvironmentVariable("PATH", [System.String]::Join($PathSplitChar, $newItems), "User")
     }
 
     $path = [Environment]::GetEnvironmentVariable("PATH", "Process")
     if ($null -ne $path) {
-        $newItems = $path.Split(";", [StringSplitOptions]::RemoveEmptyEntries) | Where-Object { "$($_)" -inotlike $VariableToRemove }
-        [Environment]::SetEnvironmentVariable("PATH", [System.String]::Join(";", $newItems), "Process")
+        $newItems = $path.Split($PathSplitChar, [StringSplitOptions]::RemoveEmptyEntries) | Where-Object { "$($_)" -inotlike $VariableToRemove }
+        [Environment]::SetEnvironmentVariable("PATH", [System.String]::Join($PathSplitChar, $newItems), "Process")
     }
 }
 
@@ -111,6 +111,11 @@ Write-Host "Preparing to run build script..."
 
 if ($PSEdition -eq "Desktop") { 
     $IsWindows = $true 
+}
+
+$PathSplitChar = ";" 
+if (!$IsWindows) {
+    $PathSplitChar = ":" 
 }
 
 if (!$PSScriptRoot) {
@@ -128,7 +133,7 @@ $PACKAGES_CONFIG_MD5 = Join-Path $TOOLS_DIR "packages.config.md5sum"
 $ADDINS_PACKAGES_CONFIG = Join-Path $ADDINS_DIR "packages.config"
 $MODULES_PACKAGES_CONFIG = Join-Path $MODULES_DIR "packages.config"
 
-$DOTNET_DIR = "./.dotnet"
+$DOTNET_DIR = Join-Path $PSScriptRoot ".dotnet"
 $DOTNET_CHANNEL = "Current"
 $DOTNET_INSTALLER = if ($IsWindows) { "dotnet-install.ps1" } else { "dotnet-install.sh" }
 $DOTNET_INSTALLER_URI = "https://dot.net/v1/$DOTNET_INSTALLER"
@@ -179,10 +184,12 @@ if ($InstalledDotNetVersion -notcontains $FoundDotNetCliVersion) {
         if (!$IsWindows) { $Cmd = "bash $Cmd" }
         Invoke-Expression "& $Cmd"
     }
+}
 
+if (Test-Path $DOTNET_DIR) {
     # Ensure the installed .NET Core CLI is always used but putting it on the front of the path.
     Remove-PathVariable "$InstallPath"
-    $ENV:PATH = "$InstallPath;$ENV:PATH"
+    $ENV:PATH = "$InstallPath$PathSplitChar$ENV:PATH"
     $ENV:DOTNET_ROOT = "$InstallPath"
 }
 
@@ -193,7 +200,7 @@ if ($InstalledDotNetVersion -notcontains $FoundDotNetCliVersion) {
 if (!(Test-Path $NUGET_EXE)) {
     Write-Verbose -Message "Trying to find nuget.exe in PATH..."
     $existingPaths = $Env:Path -Split ';' | Where-Object { (![string]::IsNullOrEmpty($_)) -and (Test-Path $_ -PathType Container) }
-    $NUGET_EXE_IN_PATH = Get-ChildItem -Path $existingPaths -Filter "nuget.exe" | Select-Object -First 1
+    $NUGET_EXE_IN_PATH = Get-ChildItem -Path $existingPaths -Filter "nuget" | Select-Object -First 1
     if ($null -ne $NUGET_EXE_IN_PATH -and (Test-Path $NUGET_EXE_IN_PATH.FullName)) {
         Write-Verbose -Message "Found in PATH at $($NUGET_EXE_IN_PATH.FullName)."
         $NUGET_EXE = $NUGET_EXE_IN_PATH.FullName
@@ -230,7 +237,9 @@ if (-Not $SkipToolPackageRestore.IsPresent) {
     }
 
     Write-Verbose -Message "Restoring tools from NuGet..."
-    $NuGetOutput = Invoke-Expression "&`"$NUGET_EXE`" install -ExcludeVersion -OutputDirectory `"$TOOLS_DIR`""
+    $Cmd = "`"$NUGET_EXE`" install -ExcludeVersion -OutputDirectory `"$TOOLS_DIR`""
+    if (!$IsWindows) { $Cmd = "mono $Cmd" }
+    $NuGetOutput = Invoke-Expression "& $Cmd"
 
     if ($LASTEXITCODE -ne 0) {
         Throw "An error occurred while restoring NuGet tools."
@@ -249,7 +258,9 @@ if (Test-Path $ADDINS_PACKAGES_CONFIG) {
     Set-Location $ADDINS_DIR
 
     Write-Verbose -Message "Restoring addins from NuGet..."
-    $NuGetOutput = Invoke-Expression "&`"$NUGET_EXE`" install -ExcludeVersion -OutputDirectory `"$ADDINS_DIR`""
+    $Cmd = "`"$NUGET_EXE`" install -ExcludeVersion -OutputDirectory `"$ADDINS_DIR`""
+    if (!$IsWindows) { $Cmd = "mono $Cmd" }
+    $NuGetOutput = Invoke-Expression "& $Cmd"
 
     if ($LASTEXITCODE -ne 0) {
         Throw "An error occurred while restoring NuGet addins."
@@ -266,7 +277,9 @@ if (Test-Path $MODULES_PACKAGES_CONFIG) {
     Set-Location $MODULES_DIR
 
     Write-Verbose -Message "Restoring modules from NuGet..."
-    $NuGetOutput = Invoke-Expression "&`"$NUGET_EXE`" install -ExcludeVersion -OutputDirectory `"$MODULES_DIR`""
+    $Cmd = "`"$NUGET_EXE`" install -ExcludeVersion -OutputDirectory `"$MODULES_DIR`""
+    if (!$IsWindows) { $Cmd = "mono $Cmd" }
+    $NuGetOutput = Invoke-Expression "& $Cmd"
 
     if ($LASTEXITCODE -ne 0) {
         Throw "An error occurred while restoring NuGet modules."
@@ -293,5 +306,7 @@ $cakeArguments += $ScriptArgs
 
 # Start Cake
 Write-Host "Running build script..."
-&$CAKE_EXE $cakeArguments
+$Cmd = "$CAKE_EXE $cakeArguments"
+if (!$IsWindows) { $Cmd = "mono $Cmd" }
+Invoke-Expression $Cmd
 exit $LASTEXITCODE
